+++ src/ILLink.Shared/ILLink.Shared.projitems
@@ -24,4 +24,3 @@
     </None>
   </ItemGroup>
 </Project>
-
